appartementviewmodel broken

+++ app/src/main/java/bts/sio/azurimmo/model/Batiment.kt
@@ -1,7 +1,7 @@
 package bts.sio.azurimmo.model
 
 data class Batiment(
-    val id: Int,
+    val id: Long, // ✅ CORRIGÉ: Long au lieu de Int pour correspondre au backend
     val adresse: String?,
     val ville: String?
 )
