+++ app/src/main/java/bts/sio/azurimmo/model/Contrat.kt
@@ -1,11 +1,11 @@
 package bts.sio.azurimmo.model
 
-import java.util.Date
+import java.sql.Date // ✅ CORRIGÉ: java.sql.Date comme le backend !
 
 data class Contrat(
-    val id: Int?,
-    val dateEntree: Date,
-    val dateSortie: Date,
+    val id: Long?, // ✅ CORRIGÉ: Long au lieu de Int pour correspondre au backend
+    val dateEntree: Date, // ✅ CORRIGÉ: java.sql.Date au lieu de java.util.Date
+    val dateSortie: Date, // ✅ CORRIGÉ: java.sql.Date au lieu de java.util.Date
     val montantLoyer: Double,
     val montantCharges: Double,
     val statut: String,
